+++ rest-api/src/main/java/org/madb/api/jpa/ContactRepository.java
@@ -5,8 +5,8 @@ import java.util.List;
 import org.madb.api.model.Contact;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ContactRepository extends JpaRepository<Contact, Long> {
+public interface ContactRepository extends JpaRepository<Contact, Integer> {
 	
-	List<Contact> findByProjectId(Long projectId);
+	List<Contact> findByProjectId(Integer projectId);
 
 }
